Move KernelDispatcher type and model into core.
