+++ numba_dpex/experimental/__init__.py
@@ -6,7 +6,6 @@
 yet production ready.
 """
 
-from numba.core.imputils import Registry
 
 # Temporary so that Range and NdRange work in experimental call_kernel
 from numba_dpex.core.boxing import *
@@ -20,20 +19,6 @@ from ._kernel_dpcpp_spirv_overloads import (
     _index_space_id_overloads,
     _private_array_overloads,
 )
-from .models import *
-from .types import KernelDispatcherType
-
-registry = Registry()
-lower_constant = registry.lower_constant
-
-
-@lower_constant(KernelDispatcherType)
-def dpex_dispatcher_const(context):
-    """Dummy lowerer for a KernelDispatcherType object. It is added so that a
-    KernelDispatcher can be passed as an argument to dpjit.
-    """
-    return context.get_dummy_value()
-
 
 __all__ = [
     "SPIRVKernelDispatcher",
